refactor!(core): Remove unnecessary prefixes/suffixes from filenames
Now that our headers are properly namespaced, there's no need to prefix their filenames
with gp_ or end them with _c to avoid name clashes with library users.
