+++ bindings/gprat_py.cpp
@@ -1,4 +1,4 @@
-#include "gprat/gprat_c.hpp"
+#include "gprat/gprat.hpp"
 
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
