+++ core/include/gprat/cpu/gp_algorithms.hpp
@@ -4,7 +4,7 @@
 #pragma once
 
 #include "gprat/detail/config.hpp"
-#include "gprat/gp_kernels.hpp"
+#include "gprat/kernels.hpp"
 
 #include <vector>
