+++ core/include/gprat/gpu/gp_algorithms.cuh
@@ -5,7 +5,7 @@
 
 #include "gprat/detail/config.hpp"
 
-#include "gprat/gp_kernels.hpp"
+#include "gprat/kernels.hpp"
 #include "gprat/target.hpp"
 #include <hpx/future.hpp>
 #include <vector>
