+++ core/include/gprat/gpu/gp_optimizer.cuh
@@ -5,8 +5,8 @@
 
 #include "gprat/detail/config.hpp"
 
-#include "gprat/gp_hyperparameters.hpp"
-#include "gprat/gp_kernels.hpp"
+#include "gprat/hyperparameters.hpp"
+#include "gprat/kernels.hpp"
 #include "gprat/target.hpp"
 
 #include <hpx/future.hpp>
