+++ core/include/gprat/kernels.hpp
