+++ test/src/output_correctness.cpp
@@ -1,5 +1,5 @@
-#include "gprat/gprat_c.hpp"
-#include "gprat/utils_c.hpp"
+#include "gprat/gprat.hpp"
+#include "gprat/utils.hpp"
 
 #include <catch2/catch_test_macros.hpp>
 #include <catch2/matchers/catch_matchers_floating_point.hpp>
